+++ modules/firehose-logs/main.tf
@@ -129,6 +129,7 @@ data "aws_iam_role" "existing_firehose_iam" {
 }
 
 resource "aws_iam_role" "new_firehose_iam" {
+  path               = var.iam_path
   count              = var.existing_firehose_iam != null ? 0 : 1
   tags               = local.tags
   name               = local.new_firehose_iam_name